fixed price calc on UpdateDemandOrder of non OnRecv
